reorganize gdml and add bottom tagger
Add the hand-crafted bottom tagger array, based on Figure 2.7 of the SBN
CRT tech note (DocDB 685). Also, reorganize the GDML a bit to make the
volume hierarchy shallower: now taggers have modules as daughters, rather
than having CRTPlanes which have the modules. This is necessary because
larcore has a hard-coded maximum volume recursion depth of 8, which is
exceeded with the old arrangement.