Renamed directory sbndcode/Geo into Geometry.
This follows the custom for most experiments.
A script is provided to apply changes to feature branches.
In fact, it's quite unlikely that any change is needed.
The script requires larsoft UPS package to be set up, and goes like:

    cd "${MRB_SOURCE}/sbndcode"
    bin/v06_30_00-GeoToGeometry.py -Uv

to print the changes, and after checking they make sense, add '--doit' to perform the changes.
A second application of the script should apply no change.