Adding the shielding block (overburden) back to the base gdml file
The overburden is added back to the base gdml file, but it is only available with setup "DefaultWithShielding". So the files sbnd_v02_00.gdml and sbnd_v02_00_nowires.gdml still do not have the overburden in them, as they were generated for the "Default" setup. A consequence of the addition of the overburden is that the two top CRT panels have been moved, to accomodate this ~2.83 m thick overburden.