Swap plane 0 and plane 1 in TPC 0 only
In the way the planes are currently implemented in the geometry, LArSoft cannot interpret the View correctly. Having the View set incorrectly causes problems down the road because many modules (inside and outside of sbndcode) use PlaneID and View interchangeably. This commit swaps plane 0 and plane 1 in TPC 0 only, so that View will bet correctly, but of course the planes won't be in the correct order. Further investigation in LArSoft is needed to fix the View problem, and once fixed, we will revert the planes again to match reality.